Renaming models to types, changing types to interfaces
