+++ packages/token-sdk/src/metadata/index.ts
@@ -1,5 +1,5 @@
 export * from "./coingecko-provider";
 export * from "./metadata-util";
 export * from "./metaplex-provider";
-export * from "./models";
+export * from "./types";
 export * from "./solanafm-provider";
